removed itineraryForm.js and cleaned up main.js itinerary imports
