+++ index.html
@@ -47,7 +47,9 @@
 
 
     <main>
-      <article class="save-itinerary-container"> </article>
+     
+        <button class="save-button" id="selection-save">Save Itinerary</button>
+        
       <section class="dropdowns flex-container">
         <div class="parks-Dropdown"></div>
